+++ src/pages/screenshot/nodejs.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import ScreenshotLang, { LangHead } from 'components/pages/screenshot-lang'
-import nodejs from 'components/pages/screenshot-lang/config/nodejs'
+import ScreenshotLang, { LangHead } from 'components/pages/screenshot/lang'
+import nodejs from 'components/pages/screenshot/lang/config/nodejs'
 
 const ScreenshotNodejs = () => <ScreenshotLang config={nodejs} />
